secure request

+++ pages/api/works/[id].ts
@@ -2,6 +2,9 @@ import type { NextApiRequest, NextApiResponse } from 'next'
 import dbConnect from '@/utils/mongodb';
 import WorkModel from '@/models/WorkModel';
 import { IWork } from '@/@types/work'
+import {session} from "next-auth/core/routes";
+import {getServerSession} from "next-auth";
+import {authOptions} from "@/pages/api/auth/[...nextauth]";
 
 type Data = {
 
@@ -11,13 +14,17 @@ type Data = {
 
 }
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
+    const session = await getServerSession(req,res,authOptions)
+
     const {
         query: { id },
         method,
     } = req
 
-
     if (req.method === 'GET') {
+        if(!session){
+            return res.status(401)
+        }
         try {
             dbConnect()
             const work = await WorkModel.findById(id)
@@ -29,6 +36,9 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
     }
 
     if (req.method === 'DELETE') {
+        if(!session){
+            return res.status(401)
+        }
         try {
             dbConnect()
             const work = await WorkModel.deleteOne({_id: id})
@@ -40,6 +50,9 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
     }
 
     if (req.method === 'PUT') {
+        if(!session){
+            return res.status(401)
+        }
         try {
             dbConnect()
             const work = await WorkModel.findByIdAndUpdate(id, req.body, {
@@ -52,5 +65,7 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
             return res.status(500).json({ message: 'Internal server error' })
         }
     }
+
+    return res.status(405)
 }
 export default handler
